+++ CHANGELOG.md
@@ -10,7 +10,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Custom global allocator.
 - `mash` prompt should react to exit code of last-executed process.
 - Fun BSOD-style screen on panic.
-- Fix the nasty code duplication in the `execve` and `execute_process` functions.
 - Nicer interface for ANSI colour escape codes (and coloured printing in general).
 - Get `mash` to "bundle together" stuff with quotes and generally handle quotes properly.
 
@@ -24,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 - Built-in `pwd` from `mash` (to be later implemented as a standalone program)
+- `vec_into_nix_bytes` and `vec_into_nix_strings` (obsolete)
 
 ## [0.1.0-beta.1] - 2025-05-26